% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{dpplot}
\alias{dpplot}
\title{Title Dynamic prediction plot for individual}
\usage{
dpplot(id_num, tmin = 0, tmax = 12, jmfit, ...)
}
\arguments{
\item{id_num}{individual id number in the list}

\item{tmin}{the start time}

\item{tmax}{the end time}

\item{jmfit}{the joint model objects}

\item{...}{}
}
\value{
a list of lme and surv predicts
}
\description{
Title Dynamic prediction plot for individual
}
