% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset.R
\docType{data}
\name{swan_amh0}
\alias{swan_amh0}
\title{Title: SWAN AMH numeric dataset}
\format{
A data frame with 3326 rows and 25 variables
}
\source{
https://www.swanstudy.org/
}
\usage{
data(swan_amh0)
}
\description{
A dataset containing
}
\details{
\itemize{
\item {id} \code{chr} SWAN ID
\item {age} \code{num} Age of the individual; TV - Age
\item {age0} \code{num} Age of the individual at the start of the study; Age - SCRN
\item {alcohol} \code{fct} Alcohol usage; TV - Alcohol Use;
type0="None", 1="<1/wk", 2="1-7/wk", 3=">7/wk";
\item {amh} \code{num} AMH
\item {bc_pills} \code{fct} Ever use of Birth Control Pills; TV - HORMEVER;
\item {bmi} \code{num} BMI; TV - BMI from PHYS; "BMI - PHYS at V0";  there are two of them
\item {bmi_cat} \code{cat} BMI as catergorical; 1=15-24.9999, 2=25-29.9999, 3=30-57
\item {children} \code{num} Number of children; Children - SCRN
\item {edu} \code{fct} Educational Attainment; TV - Educational Attainment - SCRN;
1 = "High School or Less", 2 = "Some College, 3 = "College degree or higher".

\item {employed} \code{fct} Employment Status; TV - Employment Status;
0 = "Unemployed",
1 = "Employed".

\item {ethnic} \code{char} Ethnicity from RACE;
"1=\"Caucasian\" 2=\"African American\" 3=\"Chinese\" 4=\"Hispanic\" 5=\"Japanese\";
\item {etime}
\item {event}
\item {fmp_age} \code{num} Age at final menstrual period; TV - Age at FMP;
\item {health} \code{num} Health; SCRN - Overall Health;
1="Excellent", 2="Very good", 3="Good", 4="Fair", 5="Poor", OTHER="Missing"
\item {married} \code{fct} Marital Status; SCRN - Marital Status;
\item {menopause} \code{num} Menopause status; Menopause status;
0='Pre', 1='Early Peri', 2='Late Peri', 3='Post'
\item {parity} \code{num} Parity; Parity - SCRN; The size of the family ???;
\item {phys_act} \code{num} Physical Activity Score; TV - Total Physical Activity Score w/o Work;
\item {site} \code{cat} Site;
11:M, 12:B, 13:C, 14:UCD, 15:UCLA, 16:NJ, 17:P
\item {smoke} \code{fct} Smoking Status; TV - Smoking Status;
1="Never smoked", 2="Past only", 3="Current smoker"
\item {start_time} \code{num} Start time; TV - Start Time;
\item {time} \code{num} Number of months between sample and FMP date; NUMBER OF MONTHS BTW SAMPLE & FMP DATE;
\item {train} \code{cat} subset of training or testing
\item {visit} \code{char} SWAN VISIT;
}
}
\keyword{datasets}
